Change graph to lines and bars instead of splines (#66)
Splines are more beautiful, but they have a few issues where sometimes the data
appears to go over a value and then come down, making the graph look wrong.

Also, I changed the "Ebola deaths" to be a bar chart instead of a line chart
because it's not a temporal value (as the cumulative ones are). We're just
showing the number of deaths per year, and there's no (obvious) relationship
between the values.